Move date parsing to separate file
